invest-engine/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Scheduler
apscheduler==3.10.4

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
lxml==4.9.3

# Date/Time
pytz==2023.3

# Logging
loguru==0.7.2

# Environment Variables
python-dotenv==1.0.0

# AI/NLP (optional, for future analysis features)
openai==1.3.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0